requirements.txt
streamlit
openai
python-dotenv
langchain-openai
langchain-core
langchain-community
duckduckgo-search
ddgs
langgraph
pydantic